requirements.txt
mode
faust
websockets
pydot
python-socketio
sanic
kombu
redis
aioredis